Added lexer, it seems to work! :D
This adds a lexer for the DOT language. The lexer only processes the raw
character stream and returns a token stream. As such, it is not the most
readable output, but it's a lot nicer to handle than the character
stream.

Based on testing on the files in the `dot-examples` directory, my lexer
seems to work correctly (apart from the missing **parsing a HTML
string**, which is likely going to be a pain to implement...).

Misc: Updated the Makefile to also clean the src/build directory;
updated README to include a todo-list; added a `lexTest` function to
`Graphics/DOT.idr` which takes a file path as a string and runs `lex` on
it.