Added 'attr_stmt', redid keywords in parser.
An 'attr_stmt' supports only 3 keywords, so I needed the keywords
separated out into individual parsers. As a nice benefit, this made the
`keyword` parser a whole lot nicer.
Keywords are also now case-insensitive by the time they get from the
Lexer to the Parser, since it's simpler to deal with.