Added edges, edgeops, and subgraphs.
This is where things get complicated, it seems: 'edge_stmt', 'edgeRHS',
'subgraph', 'stmt_list', and 'stmt' form a cycle, meaning that we need
to use a `mutual` block to define things neatly. It was a bit tricky to
figure out which helpers could go where, and how to 'edgeops' correctly,
but I think the current version is okay for now. Worst case, I can
always refactor things later.