Added graphs, actually implemented subgraphs. Nearly there...
I thought I was done, but nope, it turns out I'd forgotten all about
whitespace and comments ^^;
It also turned out I'd left a hole in the 'subgraph' parser. Whoops.
That is now implemented, along with 'graph' itself. Had to rename the
"graph" keyword AST/Token thingy for that to look the way I think makes
the most sense, but that was a quick refactor.

The todo-list has grown a bit, but there's (hopefully) almost an entire
parser there now.