Refactored whitespace to not store the actual WS.
The lexer used to store the whitespace, but I don't think we actually
need to.
More annoying is that the language reference technically specifies "any
amount of whitespace [...]", which I interpret as including "no amount
of whitespace". I tried doing a quick refactor on the Lexer, but it
seems a bit more complicated than just that, so for now, I'll just keep
the Lexer which requires at least one space. If I run into problems with
this, I'll obviously have to refactor things properly.