Refactored Lexer to discard whitespace.
After a couple of attempts at handling whitespace in the Parser, I
realised I was just throwing it away with `ignore` everywhere, so why
bother having it in the first place? I then went back and rewatched Dr
Chris Brown's lecture on lexical analysis and realised that this was
indeed exactly what a lexer is meant to do: turn the program into the
tokens that are relevant to the parsing stage, i.e. discard whitespace.
It should also, which is my next todo-item, discard comments. Neither
whitespace nor comments are relevant to the "program" from the
compiler's perspective, so ignore them as early as possible.