[ v0.2.0 ] New AST, parser, and interfaces
This updates the AST to be more restrictive as to what goes where
(previously you could put any `DOT` anywhere, e.g. a subgraph where
there should be a keyword). Along with this, the parser obviously needed
changing to produce this new AST.

Interfaces (yes, plural) for converting to the new AST types have been
added, along with `Show` and `Eq` implementations.

Hopefully this is more convenient to work with. GLHF!